Add tests for obtaining url parameter
Create the UrlParameter test file
to test the getRequestedUrl function in the
OSCALLoader, allowing us to determine
whether the method of obtaining the url
parameter works as expected.
