+++ package.json
@@ -15,8 +15,8 @@
     "@material-ui/core": "^4.11.4",
     "@material-ui/icons": "^4.11.2",
     "@material-ui/lab": "^4.0.0-alpha.57",
-    "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/dom": "^8.1.0",
+    "@testing-library/jest-dom": "^5.11.6",
     "@testing-library/react": "^11.2.2",
     "history": "^5.0.0",
     "prop-types": "^15.7.2",
