+++ src/components/OSCALComponentDefinition.test.js
@@ -3,17 +3,9 @@ import { render, screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import { OSCALComponentLoader } from "./OSCALLoader";
 import OSCALComponentDefinition from "./OSCALComponentDefinition";
-import testOSCALMetadata, {
-  testExternalComponentOSCALMetadata,
-} from "./OSCALMetadata.test";
-import testOSCALResponsibleRoles, {
-  testExternalOSCALResponsibleRoles,
-} from "./OSCALResponsibleRoles.test";
-import {
-  componentDefinitionTestData,
-  externalComponentDefinitionTestData,
-} from "../test-data/ComponentsData";
-import { componentDefinitionUrl } from "../test-data/Urls";
+import testOSCALMetadata from "./OSCALMetadata.test";
+import testOSCALResponsibleRoles from "./OSCALResponsibleRoles.test";
+import { componentDefinitionTestData } from "../test-data/ComponentsData";
 
 test("OSCALComponentDefinition loads", () => {
   render(<OSCALComponentLoader />);
@@ -23,15 +15,6 @@ function componentDefinitionRenderer() {
   render(
     <OSCALComponentDefinition
       componentDefinition={componentDefinitionTestData}
-    />
-  );
-}
-
-function externalUrlComponentDefinitionRenderer(externalUrl) {
-  render(
-    <OSCALComponentDefinition
-      componentDefinition={externalComponentDefinitionTestData}
-      parentUrl={externalUrl}
     />
   );
 }
@@ -52,61 +35,14 @@ function testOSCALComponentDefinitionComponent(parentElementName, renderer) {
   });
 }
 
-function testExternalOSCALComponentDefinitionComponent(
-  parentElementName,
-  renderer,
-  externalUrl
-) {
-  test(`${parentElementName} shows component title`, () => {
-    renderer(externalUrl);
-    const result = screen.getByText("test component 1");
-    expect(result).toBeVisible();
-  });
-  test(`${parentElementName} shows component description`, async () => {
-    renderer(externalUrl);
-    userEvent.hover(screen.getByText("test component 1"));
-    expect(
-      await screen.findByText(
-        "This is a software component that implements basic authentication mechanisms."
-      )
-    ).toBeInTheDocument();
-  });
-}
+testOSCALMetadata("OSCALComponentDefinition", componentDefinitionRenderer);
 
-function testUrlOSCALComponentDefinition(url) {
-  const externalUrl = url.searchParams.get("url");
-  if (!externalUrl) {
-    testOSCALMetadata("OSCALComponentDefinition", componentDefinitionRenderer);
+testOSCALComponentDefinitionComponent(
+  "OSCALComponentDefinition",
+  componentDefinitionRenderer
+);
 
-    testOSCALComponentDefinitionComponent(
-      "OSCALComponentDefinition",
-      componentDefinitionRenderer
-    );
-
-    testOSCALResponsibleRoles(
-      "OSCALComponentDefinition",
-      componentDefinitionRenderer
-    );
-    return;
-  }
-  testExternalComponentOSCALMetadata(
-    "ExternalOSCALComponentDefinition",
-    externalUrlComponentDefinitionRenderer,
-    externalUrl
-  );
-  testExternalOSCALComponentDefinitionComponent(
-    "ExternalOSCALComponentDefinition",
-    externalUrlComponentDefinitionRenderer,
-    externalUrl
-  );
-  testExternalOSCALResponsibleRoles(
-    "ExternalOSCALComponentDefinition",
-    externalUrlComponentDefinitionRenderer,
-    externalUrl
-  );
-}
-
-const url = new URL("https://www.oscal-profile-test.com");
-testUrlOSCALComponentDefinition(url);
-url.searchParams.set("url", componentDefinitionUrl);
-testUrlOSCALComponentDefinition(url);
+testOSCALResponsibleRoles(
+  "OSCALComponentDefinition",
+  componentDefinitionRenderer
+);
