+++ src/components/OSCALMetadata.test.js
@@ -28,84 +28,6 @@ export default function testOSCALMetadata(parentElementName, renderer) {
   });
 }
 
-export function testExternalSspOSCALMetadata(
-  parentElementName,
-  renderer,
-  externalURL
-) {
-  test(`${parentElementName} displays title`, () => {
-    renderer(externalURL);
-    const result = screen.getByRole("heading", {
-      name: "CSP IaaS System Security Plan",
-    });
-    expect(result).toBeVisible();
-  });
-
-  test(`${parentElementName} displays version`, () => {
-    renderer(externalURL);
-    const result = screen.getByText("0.1");
-    expect(result).toBeVisible();
-  });
-}
-
-export function testExternalProfileOSCALMetadata(
-  parentElementName,
-  renderer,
-  externalURL
-) {
-  test(`${parentElementName} displays title`, () => {
-    renderer(externalURL);
-    const result = screen.getByRole("heading", {
-      name: "NIST Special Publication 800-53 Revision 4 LOW IMPACT BASELINE",
-    });
-    expect(result).toBeVisible();
-  });
-
-  test(`${parentElementName} displays version`, () => {
-    renderer(externalURL);
-    const result = screen.getByText("2015-01-22");
-    expect(result).toBeVisible();
-  });
-
-  test(`${parentElementName} displays parties`, () => {
-    renderer(externalURL);
-    const partiesElement = screen.getByText("Parties").closest("ul");
-    const result = within(partiesElement).getByText(
-      "Joint Task Force, Transformation Initiative"
-    );
-    expect(result).toBeVisible();
-  });
-}
-
-export function testExternalComponentOSCALMetadata(
-  parentElementName,
-  renderer,
-  externalURL
-) {
-  test(`${parentElementName} displays title`, () => {
-    renderer(externalURL);
-    const result = screen.getByRole("heading", {
-      name: "Test Component Definition",
-    });
-    expect(result).toBeVisible();
-  });
-
-  test(`${parentElementName} displays version`, () => {
-    renderer(externalURL);
-    const result = screen.getByText("External Component Definition");
-    expect(result).toBeVisible();
-  });
-
-  test(`${parentElementName} displays parties`, () => {
-    renderer(externalURL);
-    const partiesElement = screen.getByText("Parties").closest("ul");
-    const result = within(partiesElement).getByText(
-      "Test Vendor External Component"
-    );
-    expect(result).toBeVisible();
-  });
-}
-
 if (!require.main) {
   testOSCALMetadata("OSCALMetadata", metadataRenderer);
 }
