+++ src/components/OSCALProfile.test.js
@@ -2,20 +2,10 @@ import React from "react";
 import { render, screen, fireEvent, act } from "@testing-library/react";
 import { OSCALProfileLoader } from "./OSCALLoader";
 import OSCALProfile from "./OSCALProfile";
-import testOSCALMetadata, {
-  testExternalProfileOSCALMetadata,
-} from "./OSCALMetadata.test";
-import testOSCALBackMatter, {
-  testExternalProfileOSCALBackMatter,
-} from "./OSCALBackMatter.test";
-import {
-  parentUrlTestData,
-  rev4LowBaselineProfileJson,
-} from "../test-data/Urls";
-import {
-  externalProfileTestData,
-  profileTestData,
-} from "../test-data/CommonData";
+import testOSCALMetadata from "./OSCALMetadata.test";
+import testOSCALBackMatter from "./OSCALBackMatter.test";
+import { parentUrlTestData } from "../test-data/Urls";
+import { profileTestData } from "../test-data/CommonData";
 
 test("OSCALProfile loads", () => {
   render(<OSCALProfileLoader />);
@@ -23,16 +13,7 @@ test("OSCALProfile loads", () => {
 
 function profileRenderer() {
   render(
-    <OSCALProfile
-      profile={profileTestData}
-      parentUrl={parentUrlTestData}
-    />
-  );
-}
-
-function externalUrlProfileRenderer(externalUrl) {
-  render(
-    <OSCALProfile profile={externalProfileTestData} parentUrl={externalUrl} />
+    <OSCALProfile profile={profileTestData} parentUrl={parentUrlTestData} />
   );
 }
 
@@ -68,54 +49,8 @@ function testOSCALProfile(parentElementName, renderer) {
   });
 }
 
-function testExternalOSCALProfile(parentElementName, renderer, externalUrl) {
-  test(`${parentElementName} displays controls`, async () => {
-    act(() => {
-      renderer(externalUrl);
-    });
-    const result = await screen.findByText("ca-1", { timeout: 10000 });
-    expect(result).toBeVisible();
-  });
+testOSCALMetadata("OSCALProfile", profileRenderer);
 
-  test(`${parentElementName} displays control modifications`, async () => {
-    renderer(externalUrl);
-    const modButton = await screen.findByRole(
-      "button",
-      { name: "ca-3 modifications" },
-      { timeout: 10000 }
-    );
-    fireEvent.click(modButton);
-    expect(await screen.findByText("Modifications")).toBeVisible();
-    expect(await screen.findByText("Adds")).toBeVisible();
-  });
-}
+testOSCALProfile("OSCALProfile", profileRenderer);
 
-function testUrlOSCALProfile(url) {
-  const externalUrl = url.searchParams.get("url");
-  if (!externalUrl) {
-    testOSCALMetadata("OSCALProfile", profileRenderer);
-    testOSCALProfile("OSCALProfile", profileRenderer);
-    testOSCALBackMatter("OSCALProfile", profileRenderer);
-    return;
-  }
-  testExternalProfileOSCALMetadata(
-    "ExternalOSCALProfile",
-    externalUrlProfileRenderer,
-    externalUrl
-  );
-  testExternalOSCALProfile(
-    "ExternalOSCALProfile",
-    externalUrlProfileRenderer,
-    externalUrl
-  );
-  testExternalProfileOSCALBackMatter(
-    "ExternalOSCALProfile",
-    externalUrlProfileRenderer,
-    externalUrl
-  );
-}
-
-const url = new URL("https://www.oscal-profile-test.com");
-testUrlOSCALProfile(url);
-url.searchParams.set("url", rev4LowBaselineProfileJson);
-testUrlOSCALProfile(url);
+testOSCALBackMatter("OSCALProfile", profileRenderer);
